Manage LegalDisplaymer y AppFrame parent
